rename: hook 또는 함수의 파라미터 타입 ~Props -> ~Params
- 함수형 컴포넌트의 파라미터 타입을 ~Props로 묶어 정의하던 걸 hook과 일반 함수에도 적용하면서 ~Props 타입 네이밍으로 사용하고 있었다.
- 그러나 hook과 일반 함수에서 ~Props는 의미적으로 어색하기 때문에 파라미터를 의미하는 ~Params로 이름을 수정하였다.
